// src/lib/types.ts
export type EntityId = string | number;

export type Company = {
  id: EntityId;
  code: string;
  name: string;
  document: string;
  nickname: string;
  lastProcess: string;
  mode?: "Customizada" | "Simples";
  taxation: TaxationType;
  userId?: string;
};

export type TaxationType = "Lucro Real" | "Lucro Presumido" | "Simples Nacional" | "Imunes/Isentas";
// ...
export type AccountPlanEntry = {
  id: string | number;
  docId?: string;
  account: string;
  reducedAccount?: string;
  synthetic: string;
  classificationCode?: string;
  classification: string;
  nickname: string;
};